Allow a semi arbitrary # of frames. Chunk frames into batches.
Chunk frames into batches controlled by frame_stride.  Increase neighbor_stride by 5 because it makes the transitions between the chunks smoother.

There are some artifacts generated with this process, but, it is usable.

The point of this is to not load all of the frames into VRAM at the same time thereby making it work on people with graphics cards with less VRAM and on an arbitrary length video.